+++ server/pipeline/src/main/java/org/bithon/server/pipeline/metrics/exporter/MetricMessageHandler.java
@@ -21,7 +21,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.bithon.component.commons.concurrency.NamedThreadFactory;
 import org.bithon.component.commons.utils.CollectionUtils;
 import org.bithon.component.commons.utils.SupplierUtils;
-import org.bithon.server.pipeline.common.transform.TransformSpec;
+import org.bithon.server.pipeline.common.transformer.TransformSpec;
 import org.bithon.server.pipeline.metrics.MetricPipelineConfig;
 import org.bithon.server.pipeline.metrics.MetricsAggregator;
 import org.bithon.server.pipeline.metrics.topo.ITopoTransformer;
