+++ server/pipeline/src/main/java/org/bithon/server/pipeline/metrics/input/IMetricInputSource.java
@@ -20,7 +20,7 @@ import com.fasterxml.jackson.annotation.JsonTypeInfo;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import org.bithon.server.pipeline.common.transform.TransformSpec;
+import org.bithon.server.pipeline.common.transformer.TransformSpec;
 import org.bithon.server.storage.datasource.ISchema;
 
 import java.time.Duration;
