+++ server/pipeline/src/main/java/org/bithon/server/pipeline/metrics/transform/ConnectionStringTransformer.java
@@ -21,8 +21,8 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonTypeName;
 import lombok.Getter;
 import org.bithon.server.commons.utils.DbUtils;
-import org.bithon.server.pipeline.common.transform.transformer.ITransformer;
-import org.bithon.server.pipeline.common.transform.transformer.TransformResult;
+import org.bithon.server.pipeline.common.transformer.ITransformer;
+import org.bithon.server.pipeline.common.transformer.TransformResult;
 import org.bithon.server.storage.datasource.input.IInputRow;
 
 /**
