+++ server/pipeline/src/main/java/org/bithon/server/pipeline/metrics/transform/UriNormalizationTransformer.java
@@ -24,8 +24,8 @@ import com.fasterxml.jackson.annotation.OptBoolean;
 import lombok.Getter;
 import org.bithon.component.commons.utils.StringUtils;
 import org.bithon.server.pipeline.common.service.UriNormalizer;
-import org.bithon.server.pipeline.common.transform.transformer.ITransformer;
-import org.bithon.server.pipeline.common.transform.transformer.TransformResult;
+import org.bithon.server.pipeline.common.transformer.ITransformer;
+import org.bithon.server.pipeline.common.transformer.TransformResult;
 import org.bithon.server.storage.datasource.input.IInputRow;
 
 /**
