+++ server/pipeline/src/test/java/org/bithon/server/pipeline/tracing/transform/UrlSanitizeTransformTest.java
@@ -17,7 +17,7 @@
 package org.bithon.server.pipeline.tracing.transform;
 
 import com.google.common.collect.ImmutableMap;
-import org.bithon.server.pipeline.common.transform.transformer.AbstractTransformer;
+import org.bithon.server.pipeline.common.transformer.AbstractTransformer;
 import org.bithon.server.pipeline.tracing.transform.sanitization.UrlSanitizeTransformer;
 import org.bithon.server.storage.tracing.TraceSpan;
 import org.junit.Assert;
